+++ tool/script/zlib_install.sh
@@ -2,15 +2,20 @@
 
 ## https://www.zlib.net/
 
-zip_dir="${PWD}/zlib1211.zip"
+if [ $# != 2 ]; then
+    printf "USAGE (sudo permission may be needed):\n"
+	printf "	\`${0} <ZIP_DIR> <INSTALL_DIR>\`\n"
+	exit 1;
+fi
+
+zip_dir="${1}"
+ins_dir="${2}"
 new_folder="${PWD}/zlib_source"
-HIKSDK_DIR="/data1/duruyao/HikSDK"
 
 mkdir -p ${new_folder} && rm -rf ${new_folder}/*
 unzip -q ${zip_dir} -d ${new_folder}
 
 src_dir="${new_folder}/`ls ${new_folder}`"
-ins_dir="${HIKSDK_DIR}/zlib"
 
 mkdir -p ${ins_dir} && rm -rf ${ins_dir}/*
 cd ${src_dir}
